- Corrected ambiguous calling of pow in hLinearFitPolynomialX
git-svn-id: http://usg.lofar.org/svn/code/trunk@4902 e3a606d5-b720-0410-b6ad-cb58f335abf8

+++ src/CR-Tools/implement/Pypeline/hftools.cc
@@ -5124,7 +5124,7 @@ void HFPP_FUNC_NAME(
       *itout=(*itx) * (*itx);
       break;
     default:
-      *itout=pow(*itx,*itn);
+      *itout=pow(*itx,(int)*itn);
       break;
     };
     ++itout; ++itn;
